SAGE-1581: correct false compute unit matches; match entire compute mac address
Before this change an empty ("") compute 'serial_no' (i.e. mac address) was
matching as the empty ("") string is "in" every other string. Instead,
only match the compute if the 'serial_no' (i.e. mac address) is an exact
match to the mac address embedded in the kube node name.